Update client and channel providing
