+++ core/src/main/java/com/rettichlp/unicacityaddon/base/teamspeak/TeamSpeakAuthenticator.java
@@ -29,102 +29,99 @@ import java.util.ArrayList;
 import java.util.List;
 
 /**
- * The original code is available at: <a href="https://github.com/labymod-addons/teamspeak">https://github.com/labymod-addons/teamspeak</a>.
+ * This code was modified. The original code is available at: <a href="https://github.com/labymod-addons/teamspeak">https://github.com/labymod-addons/teamspeak</a>.
  * <p>
  * The following code is subject to the LGPL Version 2.1.
  *
  * @author jumpingpxl
+ * @author RettichLP
  */
 public class TeamSpeakAuthenticator {
 
-  private final DefaultTeamSpeakAPI teamSpeakAPI;
-  private final List<Path> possibleDirectories;
+    private final TeamSpeakAPI teamSpeakAPI;
+    private final List<Path> possibleDirectories;
 
-  public TeamSpeakAuthenticator(UnicacityAddon unicacityAddon, DefaultTeamSpeakAPI teamSpeakAPI) {
-    this.teamSpeakAPI = teamSpeakAPI;
-    this.possibleDirectories = new ArrayList<>();
+    public TeamSpeakAuthenticator(UnicacityAddon unicacityAddon, TeamSpeakAPI teamSpeakAPI) {
+        this.teamSpeakAPI = teamSpeakAPI;
+        this.possibleDirectories = new ArrayList<>();
 
-    OperatingSystem platform = OperatingSystem.getPlatform();
-    if (platform == OperatingSystem.WINDOWS) {
-      this.loadWindowsDirectories();
-    } else if (platform == OperatingSystem.MACOS) {
-      this.loadMacOsDirectories();
-    } else if (platform == OperatingSystem.LINUX) {
-      this.loadLinuxDirectories();
-    } else {
-      unicacityAddon.logger().warn("Cannot automatically resolve the Auth Key on " + platform + "!");
-    }
-  }
-
-  public boolean authenticate() {
-    boolean triedAuthenticating = false;
-    for (Path possibleDirectory : this.possibleDirectories) {
-      Path clientQueryPath = possibleDirectory.resolve("clientquery.ini");
-      if (!IOUtil.exists(clientQueryPath)) {
-        continue;
-      }
-
-      try {
-        String contents = new String(Files.readAllBytes(clientQueryPath), StandardCharsets.UTF_8);
-        String[] split = contents.split("\n");
-        for (String line : split) {
-          line = line.trim();
-          if (!line.startsWith("api_key=")) {
-            continue;
-          }
-
-          String apiKey = line.substring("api_key=".length()).replace("\r", "");
-          this.teamSpeakAPI.authenticate(apiKey);
-          triedAuthenticating = true;
+        OperatingSystem platform = OperatingSystem.getPlatform();
+        if (platform == OperatingSystem.WINDOWS) {
+            this.loadWindowsDirectories();
+        } else if (platform == OperatingSystem.MACOS) {
+            this.loadMacOsDirectories();
+        } else if (platform == OperatingSystem.LINUX) {
+            this.loadLinuxDirectories();
+        } else {
+            unicacityAddon.logger().warn("Cannot automatically resolve the Auth Key on " + platform + "!");
         }
-      } catch (IOException e) {
-        e.printStackTrace();
-      }
     }
 
-    return triedAuthenticating;
-  }
+    public void authenticate() {
+        for (Path possibleDirectory : this.possibleDirectories) {
+            Path clientQueryPath = possibleDirectory.resolve("clientquery.ini");
+            if (!IOUtil.exists(clientQueryPath)) {
+                continue;
+            }
 
-  private void loadWindowsDirectories() {
-    Path appData = Paths.get(System.getenv("AppData"));
-    Path local = appData.getParent().resolve("Local");
+            try {
+                String contents = new String(Files.readAllBytes(clientQueryPath), StandardCharsets.UTF_8);
+                String[] split = contents.split("\n");
+                for (String line : split) {
+                    line = line.trim();
+                    if (!line.startsWith("api_key=")) {
+                        continue;
+                    }
 
-    this.checkForPossibleDirectories(
-        appData.resolve("TS3Client"),
-        local.resolve("TS3Client")
-    );
-  }
-
-  private void loadMacOsDirectories() {
-    Path userHome = Paths.get(System.getProperty("user.home"));
-    Path library = userHome.resolve("Library");
-    Path applicationSupport = library.resolve("Application Support");
-
-    this.checkForPossibleDirectories(
-        applicationSupport.resolve("TeamSpeak 3")
-    );
-  }
-
-  private void loadLinuxDirectories() {
-    Path userHome = Paths.get(System.getProperty("user.home"));
-
-    this.checkForPossibleDirectories(
-        userHome.resolve(".ts3client")
-    );
-  }
-
-  private void checkForPossibleDirectories(Path... paths) {
-    for (Path path : paths) {
-      if (!IOUtil.exists(path)) {
-        continue;
-      }
-
-      this.possibleDirectories.add(path);
-      Path config = path.resolve("config");
-      if (IOUtil.exists(config)) {
-        this.possibleDirectories.add(config);
-      }
+                    String apiKey = line.substring("api_key=".length()).replace("\r", "");
+                    this.teamSpeakAPI.authenticate(apiKey);
+                }
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        }
     }
-  }
+
+    private void loadWindowsDirectories() {
+        Path appData = Paths.get(System.getenv("AppData"));
+        Path local = appData.getParent().resolve("Local");
+
+        this.checkForPossibleDirectories(
+                appData.resolve("TS3Client"),
+                local.resolve("TS3Client")
+        );
+    }
+
+    private void loadMacOsDirectories() {
+        Path userHome = Paths.get(System.getProperty("user.home"));
+        Path library = userHome.resolve("Library");
+        Path applicationSupport = library.resolve("Application Support");
+
+        this.checkForPossibleDirectories(
+                applicationSupport.resolve("TeamSpeak 3")
+        );
+    }
+
+    private void loadLinuxDirectories() {
+        Path userHome = Paths.get(System.getProperty("user.home"));
+
+        this.checkForPossibleDirectories(
+                userHome.resolve(".ts3client")
+        );
+    }
+
+    private void checkForPossibleDirectories(Path... paths) {
+        for (Path path : paths) {
+            if (!IOUtil.exists(path)) {
+                continue;
+            }
+
+            this.possibleDirectories.add(path);
+            Path config = path.resolve("config");
+            if (IOUtil.exists(config)) {
+                this.possibleDirectories.add(config);
+            }
+        }
+    }
 }
