+++ core/src/main/java/com/rettichlp/unicacityaddon/base/teamspeak/listener/ClientLeftViewListener.java
@@ -16,50 +16,35 @@
 
 package com.rettichlp.unicacityaddon.base.teamspeak.listener;
 
-import com.rettichlp.unicacityaddon.base.teamspeak.DefaultTeamSpeakAPI;
-import com.rettichlp.unicacityaddon.base.teamspeak.models.DefaultChannel;
-import com.rettichlp.unicacityaddon.base.teamspeak.models.DefaultServer;
-import com.rettichlp.unicacityaddon.base.teamspeak.models.DefaultUser;
+import com.rettichlp.unicacityaddon.base.teamspeak.TeamSpeakAPI;
+import com.rettichlp.unicacityaddon.base.teamspeak.models.Server;
 
 /**
- * The original code is available at: <a href="https://github.com/labymod-addons/teamspeak">https://github.com/labymod-addons/teamspeak</a>.
+ * This code was modified. The original code is available at: <a href="https://github.com/labymod-addons/teamspeak">https://github.com/labymod-addons/teamspeak</a>.
  * <p>
  * The following code is subject to the LGPL Version 2.1.
  *
  * @author jumpingpxl
+ * @author RettichLP
  */
-public class ClientLeftViewListener extends DefaultListener {
+public class ClientLeftViewListener extends Listener {
 
-  public ClientLeftViewListener() {
-    super("notifyclientleftview");
-  }
-
-  @Override
-  public void execute(DefaultTeamSpeakAPI teamSpeakAPI, String[] args) {
-    Integer schandlerId = this.get(args, "schandlerid", Integer.class);
-    if (schandlerId == null) {
-      return;
+    public ClientLeftViewListener() {
+        super("notifyclientleftview");
     }
 
-    DefaultServer selectedServer = teamSpeakAPI.getSelectedServer();
-    if (selectedServer == null || selectedServer.getId() != schandlerId) {
-      return;
-    }
+    @Override
+    public void execute(TeamSpeakAPI teamSpeakAPI, String[] args) {
+        Integer schandlerId = this.get(args, "schandlerid", Integer.class);
+        if (schandlerId == null) {
+            return;
+        }
 
-    DefaultChannel selectedChannel = selectedServer.getSelectedChannel();
-    Integer channelId = this.get(args, "cfid", Integer.class);
-    if (channelId == null || selectedChannel == null || selectedChannel.getId() != channelId) {
-      return;
-    }
+        Server selectedServer = teamSpeakAPI.getSelectedServer();
+        if (selectedServer == null || selectedServer.getId() != schandlerId) {
+            return;
+        }
 
-    Integer clientId = this.get(args, "clid", Integer.class);
-    if (clientId == null) {
-      return;
+        teamSpeakAPI.controller().refreshCurrentServer(schandlerId);
     }
-
-    DefaultUser user = selectedChannel.getUser(clientId);
-    if (user != null) {
-      selectedChannel.getUsers().remove(user);
-    }
-  }
 }
