--- core/src/main/java/com/rettichlp/unicacityaddon/base/teamspeak/querycommand/Channel.java
@@ -1,14 +1,0 @@
-package com.rettichlp.unicacityaddon.base.teamspeak.querycommand;
-
-import lombok.Data;
-
-/**
- * @author RettichLP
- */
-@Data
-public class Channel {
-
-    private final int id;
-    private final String name;
-    private final int pid;
-}
