--- core/src/main/java/com/rettichlp/unicacityaddon/base/teamspeak/querycommand/Response.java
@@ -1,4 +1,0 @@
-package com.rettichlp.unicacityaddon.base.teamspeak.querycommand;
-
-public class Response {
-}
